Add a unit testing framework for direct translations
The idea is to pull existing example Apidoc definitions (though we can also
manually add service.json files to the directory too) and map those to
an expected swagger.json. If either changes, we simply edit the JSON on
either end and the test doesn't change.

Please note that the existing swagger.json examples are not valid Swagger
2.0 documents; I will follow up this commit with more commits to address
those problems.